all: expose EndianMagicU32 on D-Bus for detecting endianness issues
In NetworkManager's D-Bus API, some IPv4 addresses are exposed in
network order (big endian). Note that in the meantime, all these
properties are either entirely deprecated, or the API was extended
to expose the values also as strings, agnostic to byte ordering.

Anyway. Add a property that allows to detect the servers endianness of the
server. The server will always set this property to the magic value
0x12345678 in network order.

On the client side, you can detect byte ordering issues via:

    dbus_endian_magic_u32 = g_variant_get_uint32 (variant);
    if (dbus_endian_magic_u32 != htonl (NM_ENDIAN_MAGIC_U32)) {
        /* values in network-order got scrambled. Needs correction. */
    }

https://bugzilla.redhat.com/show_bug.cgi?id=1153559
https://bugzilla.redhat.com/show_bug.cgi?id=1584584