squash! all: expose EndianMagicU32 on D-Bus for detecting endianness issues
>>>==========================================================
Rethinking, maybe this is better than the NM_ENDIAN_MAGIC_U32
because the entire feature is rather obscure and only needed
in very specific circumstances. While EndianMagicU32 seems
slightly more convenient to use (to me), it has more overhead
by adding a new D-Bus property.
<<<==========================================================

all: expose EndianMagicU32 on D-Bus for detecting endianness issues

In NetworkManager's D-Bus API, some IPv4 addresses are exposed in
network order (big endian). Note that in the meantime, all these
properties are either deprecated for good, or the API was extended
to expose the values also as strings and agnostic to byte ordering.

Anyway. Add a capability that allows to detect the servers endianness of the
server.

https://bugzilla.redhat.com/show_bug.cgi?id=1153559
https://bugzilla.redhat.com/show_bug.cgi?id=1584584