libnm/keyfile: use NMMetaSettingInfo for indexing keyfile vtable
We have NMMetaSettingType enum, which is an enum of all setting types.
We also have an efficient way to get the enum (and its NMMetaSettingInfo)
from an NMSetting, setting-name and GType.

No longer maintain the vtable for keyfile by "setting-name". Instead,
index it by NMMetaSettingType enum.

That way, we get efficient lookup, and don't need to duplicate the
functionality of finding the vtable entry for a setting.