libnm/crypto: minor cleanup confusing comment in crypto_decrypt_openssl_private_key_data()
the comment and code made it sound like parse_old_openssl_key_file() would
set @key_type if the parsing was only done partially. That is not the case,
@key_type is only set, if parsing was successful. Adjust the code.

While at it, don't require the caller to initialize @out_key_type. It's
just an enum, if we care to always set it, just do it.