libnm: fix race in nm-setting-x802-1x's setting private key functions
Do not first load the file during nm_crypto_verify_private_key(),
and later re-load it, in case we are setting a blob. Instead,
ensure we only load the file once.

This fixes a race, and also the very wrong assertion:

    priv->phase2_private_key = nm_crypto_read_file (value, NULL);
    nm_assert (priv->phase2_private_key);

We should never assert that an IO operation succeeds.

Also, we encode blobs, paths, and pkcs11 URIs all inside a binary
field. Unfortunately, there is no defined prefix for blobs (TODO).
That means, if you have a blob that happens to start with "file://"
it cannot be expressed. At least, check that the binary field that
we are setting gets detected as correct scheme type.